Spiking out some big methods into responders
Such WIP. Wow.
