+++ Gemfile
@@ -36,6 +36,7 @@ gem 'rack-rewrite', '1.2.1'
 gem 'rack-ssl-enforcer', '0.2.4'
 gem 'rails', '4.1.5'
 gem 'rails_admin', '~> 0.6.2'
+gem 'responders'
 gem 'ruby-freshbooks', '0.4.0'
 gem 'stripe', github: 'stripe/stripe-ruby'
 gem 'stripe_event'
